Plugins: Move formatter cache loading out of Sortable Table (#6978)
